+++ groovyConsole/history.md
@@ -5,6 +5,7 @@
 * Changed 'freeplaneVersionFrom' to a lower number, because I realized it worked in v1.9 versions too
 * Added a command to call groovyConsole without restrictions
 * Added button to reload script from disk.
+* Now it can handle absolute and relative links to script files
 
 ## v0.0.4